+++ src/sdhlt/sdHLCSG/csg.h
@@ -34,6 +34,7 @@
 #define DEFAULT_BRUSH_UNION_THRESHOLD 0.0f
 #define DEFAULT_TINY_THRESHOLD        0.0
 #define DEFAULT_NOCLIP      false
+#define DEFAULT_CONVEXFIX   true
 #define DEFAULT_ONLYENTS    false
 #define DEFAULT_WADTEXTURES true
 #define DEFAULT_SKYCLIP     true
@@ -284,6 +285,7 @@ extern bool     g_chart;
 extern bool     g_deterministic;
 extern bool     g_onlyents;
 extern bool     g_noclip;
+extern bool     g_convexfix;
 extern bool     g_wadtextures;
 extern bool     g_skyclip;
 extern bool     g_estimate;